+++ src/components/component.js
@@ -334,7 +334,7 @@ function setupProperty(classDef, prop, propertyDefHash) {
   classDef[name] = newDef;
 }
 
-layerUI.registerComponent = function registerComponent(tagName, classDef, force) {
+module.exports =  function registerComponent(tagName, classDef, force) {
   if (layerUI.settings.customComponents.indexOf(tagName) !== -1 && !force) return;
 
   // Insure property exists
@@ -700,6 +700,7 @@ layerUI.registerComponent = function registerComponent(tagName, classDef, force)
    *                    the default action using `evt.preventDefault()` (perhaps an event listener wanted to handle the action itself)
    */
   classDef.trigger = {
+    writable: true, // jasmine.spyOn support
     value: function trigger(eventName, details) {
       const evt = new CustomEvent(eventName, {
         detail: details,
@@ -782,5 +783,3 @@ layerUI.registerComponent = function registerComponent(tagName, classDef, force)
    */
   layerUI.components[tagName].properties = props;
 };
-
-module.exports = layerUI.registerComponent;